requirements.txt
customtkinter==5.0.1
darkdetect==0.7.1
python-crontab==2.6.0
python-dateutil==2.8.2
six==1.16.0
